requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-search-documents
openai
dotenv
azure-common
azure-core
azure-search==1.0.0b2
azure-search-documents==11.5.2
aiohttp
azure-identitiy
